+++ packages/apollo-link-firebase-firestore/CHANGELOG.md
@@ -8,6 +8,7 @@ and adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 ## [Unreleased]
 
 - updated dependencies
+- adding common QueryError type
 
 ## [1.0.1] - 2019-07-07